requirements.txt
confluent-kafka
attrs
